+++ Cargo.toml
@@ -32,7 +32,7 @@ async-compression = { version = "0.4.8", features = [
     "bzip2",
     "zstd",
 ] }
-async-fd-lock = "0.2.0"
+fs4 = "0.9.1"
 async-trait = "0.1.80"
 axum = { version = "0.7.5", default-features = false, features = [
     "tokio",
@@ -114,7 +114,7 @@ retry-policies = { version = "0.4.0", default-features = false }
 rmp-serde = { version = "1.2.0" }
 rstest = { version = "0.21.0" }
 rstest_reuse = "0.7.0"
-simd-json = { version = "0.13.10" , features = ["serde_impl"]}
+simd-json = { version = "0.13.10", features = ["serde_impl"] }
 serde = { version = "1.0.198" }
 serde_json = { version = "1.0.116" }
 serde_repr = "0.1"
