+++ crates/rattler_lock/src/utils/serde/pep440_map_or_vec.rs
@@ -35,6 +35,7 @@ impl<'de> DeserializeAs<'de, Vec<Requirement>> for Pep440MapOrVec {
                         } else {
                             Some(VersionOrUrl::VersionSpecifier(spec))
                         },
+                        #[allow(clippy::default_trait_access)]
                         marker: Default::default(),
                         origin: None,
                     })
