+++ crates/rattler_repodata_gateway/src/gateway/direct_url_query.rs
@@ -1,7 +1,8 @@
 use std::{future::IntoFuture, sync::Arc};
 
 use futures::FutureExt;
-use rattler_cache::package_cache::{CacheKey, PackageCache, PackageCacheError};
+use rattler_cache::package_cache::CacheKey;
+use rattler_cache::package_cache::{PackageCache, PackageCacheError};
 use rattler_conda_types::{
     package::{ArchiveIdentifier, IndexJson, PackageFile},
     ConvertSubdirError, PackageRecord, RepoDataRecord,
